concertou alocacao da lista de blocos vazios para funciona atraves de blocos e nao posicao-- nrup terminado  mas falta testar

+++ GerenciadorDeMemoria/src/AcessoDaMemoria.java
@@ -1,16 +1,16 @@
 
 public class AcessoDaMemoria {
 	private int p;
-	private double t;
+	private int t;
 	
 	public int getP() {
 		return p;
 	}
-	public double getT() {
+	public int getT() {
 		return t;
 	}
 	
-	public AcessoDaMemoria(int p, double t){
+	public AcessoDaMemoria(int p, int t){
 		this.p = p;
 		this.t = t;
 	}
